pyproject.toml
[project]
name = "naver-blog-auto"
version = "0.2.0"
description = "네이버 블로그 자동 작성 도구"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "ollama>=0.4.0",
    "pandas>=2.2.0",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0",
    "requests>=2.32.5",
    "streamlit>=1.54.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core"]
